+++ extensions/keepassxc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # KeePassXC Extension Changelog
 
+## [1.8.0] - 2025-10-23
+
+### Added
+
+- Support for KeePassXC placeholders in copy and paste actions.
+- Process placeholders like `{TOTP}`, `{USERNAME}`, `{PASSWORD}`, `{TITLE}`, `{URL}`, and `{NOTES}` when copying or pasting entry data.
+
 ## [1.7.1] - 2025-10-14
 
 ### Fixed